Ordered components inside Webapp into folders
