+++ webapp/src/components/login/PersistentLogin.test.js
@@ -1,7 +1,7 @@
 import { render, screen, waitFor } from "@testing-library/react";
 import PersistentLogin from "./PersistentLogin";
-import useAuth from "../hooks/useAuth";
-import useRefreshToken from "../hooks/useRefreshToken";
+import useAuth from "../../hooks/useAuth";
+import useRefreshToken from "../../hooks/useRefreshToken";
 import { Outlet } from "react-router";
 import { Typography } from "@mui/material";
