+++ webapp/src/index.js
@@ -2,20 +2,20 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import reportWebVitals from './reportWebVitals';
-import styles from './index.css';
 
 import { AuthProvider } from './context/AuthContext';
 import PrivateRoute from './components/PrivateRoute';
 
 import App from './App';
-import Login from './components/Login';
+import Login from './components/login/Login';
 import SignUp from './components/AddUser';
-import GameModeSelection from './components/GameModeSelection';
-import Game from './components/Game';
+import GameModeSelection from './components/selections/GameModeSelection';
+import RoundsGame from './components/modes/RoundsGame';
+import TimeGame from './components/modes/TimeGame';
 import Home from './components/Home';
 import Layout from './components/Layout';
-import GameTopicSelection from './components/GameTopicSelection';
-import PersistentLogin from './components/PersistentLogin';
+import GameTopicSelection from './components/selections/GameTopicSelection';
+import PersistentLogin from './components/login/PersistentLogin';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
@@ -33,7 +33,8 @@ root.render(
               <Route path="/home" element={<Home />} />
               <Route path="/gamemode" element={<GameModeSelection />} />
               <Route path="/gametopic" element={<GameTopicSelection />} />
-              <Route path="/game" element={<Game />} />
+              <Route path="/roundsgame" element={<RoundsGame />} />
+              <Route path="/timegame" element={<TimeGame />} />
             </Route>
           </Route>
         </Route>
